move handeEditRecipe function to edit page
