+++ types.ts
@@ -61,9 +61,5 @@ export type AppBaseProps = {
 export interface ChildrenLayoutProps {
   children?: ReactNode;
 }
-/* TODO */
-/* export type OnAddRecipeType = {
-  onAddRecipe: (recipe: Recipe) => void;
-}; */
 
 export type OnEditRecipeType = (recipe: Recipe) => void;
