--- types/RecipeEditPage.types.ts
@@ -1,7 +1,0 @@
-import { Recipe } from '@/types';
-import { OnEditRecipeType } from '@/types';
-
-export type RecipeEditPageProps = {
-  recipes: Recipe[];
-  onEditRecipe: OnEditRecipeType;
-};
